# Conflicts:
#	Cartfile.resolved
#	Haneke.playground/playground.xcworkspace/contents.xcworkspacedata
#	Haneke.xcodeproj/project.pbxproj
#	Haneke.xcodeproj/xcshareddata/xcschemes/Haneke-iOS.xcscheme
#	Haneke.xcodeproj/xcshareddata/xcschemes/Haneke-tvOS.xcscheme
#	Haneke.xcodeproj/xcshareddata/xcschemes/HanekeDemo.xcscheme
#	Haneke/Data.swift
#	Haneke/UIButton+Haneke.swift
#	Haneke/UIImage+Haneke.swift
#	Haneke/UIView+Haneke.swift
#	HanekeDemo/AppDelegate.swift
#	HanekeTests/CacheTests.swift
#	HanekeTests/DiskCacheTests.swift
#	HanekeTests/DiskFetcherTests.swift
#	HanekeTests/UIButton+HanekeTests.swift
#	HanekeTests/UIImage+HanekeTests.swift
#	HanekeTests/XCTestCase+Test.swift

+++ Haneke/UIImage+Haneke.swift
@@ -28,9 +28,9 @@ extension UIImage {
         }
     }
     
-    func hnk_data(compressionQuality: CGFloat = 1.0) -> Data! {
+    func hnk_data(compressionQuality: Float = 1.0) -> Data! {
         let hasAlpha = self.hnk_hasAlpha()
-        let data = hasAlpha ? self.pngData() : self.jpegData(compressionQuality: compressionQuality)
+        let data = hasAlpha ? self.pngData() : self.jpegData(compressionQuality: CGFloat(compressionQuality))
         return data
     }
